// src/app/models/OrderDetails.ts
export interface OrderDetails{
    orderId?: number;
    orderDate?: string;
    totalPrice?: number;
    city?: string;
    address?: string;
    shippingDate?: string;
    creditCard?: string;
    products?: any[];
}

//    orderId?: number;
// amount?: number;
// price?: number;
// productId: number;
// orderDate?: string;
// totalPrice?: number;
// city?: string;
// address?: string;
// shippingDate?: string;
// creditCard?: number;
// cartId?: number;